feat(shell):知识图谱榜单脚本优化

+++ vova/ads/vova_bod/knowledge_graph/sqoop_export.sh
@@ -5,8 +5,72 @@ pre_date=$1
 if [ ! -n "$1" ]; then
   pre_date=$(date -d "-1 day" +%Y-%m-%d)
 fi
+sh /mnt/vova-bigdata-scripts/common/table_check.sh mlb.mlb_vova_kg_d 800 "pt='${pre_date}'"
+if [ $? -ne 0 ];then
+exit 1
+fi
+m_sql="
+drop table if exists rec_recall.ads_vova_kg_d_pre;
+drop table if exists rec_recall.ads_vova_kg_d_new;
+CREATE TABLE IF NOT EXISTS \`rec_recall\`.\`ads_vova_kg_d_new\` (
+\`id\` int(11) NOT NULL AUTO_INCREMENT,
+\`buyer_id\` int(11) NOT NULL COMMENT '用户id',
+\`bod_id_list\` mediumtext NOT NULL COMMENT '榜单id列表',
+\`score_list\` mediumtext NOT NULL COMMENT '分数列表',
+\`update_time\` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+PRIMARY KEY (\`id\`) USING BTREE,
+KEY \`buyer_id\` (\`buyer_id\`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='知识图谱-用户榜单偏好';
+CREATE TABLE IF NOT EXISTS \`rec_recall\`.\`ads_vova_kg_d\` (
+\`id\` int(11) NOT NULL AUTO_INCREMENT,
+\`buyer_id\` int(11) NOT NULL COMMENT '用户id',
+\`bod_id_list\` mediumtext NOT NULL COMMENT '榜单id列表',
+\`score_list\` mediumtext NOT NULL COMMENT '分数列表',
+\`update_time\` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+PRIMARY KEY (\`id\`) USING BTREE,
+KEY \`buyer_id\` (\`buyer_id\`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='知识图谱-用户榜单偏好';
+"
+mysql -h rec-recall.cluster-cznqgcwo1pjt.us-east-1.rds.amazonaws.com -u dwrecallwriter -pTsLdpZumzovrAvttIqnePCJhIVxZZ7bd -e "${m_sql}"
 
-sh /mnt/vova-bigdata-scripts/common/table_check.sh  ads.ads_vova_knowledge_graph_bod_goods_rank 800 "pt='${pre_date}'"
+if [ $? -ne 0 ];then
+exit 1
+fi
+
+sqoop export \
+-Dorg.apache.sqoop.export.text.dump_data_on_error=true \
+-Dsqoop.export.records.per.statement=500 \
+-Dmapreduce.map.memory.mb=8192 \
+-Dmapreduce.reduce.memory.mb=8192 \
+--connect jdbc:mysql://rec-recall.cluster-cznqgcwo1pjt.us-east-1.rds.amazonaws.com:3306/rec_recall \
+--username dwrecallwriter --password TsLdpZumzovrAvttIqnePCJhIVxZZ7bd \
+--m 1 \
+--table ads_vova_kg_d_new \
+--hcatalog-database mlb \
+--hcatalog-table mlb_vova_kg_d \
+--columns buyer_id,bod_id_list,score_list \
+--hive-partition-key pt \
+--hive-partition-value ${pre_date} \
+--fields-terminated-by '\001'
+
+if [ $? -ne 0 ];then
+exit 1
+fi
+echo "----------开始rename-------"
+mysql -h rec-recall.cluster-cznqgcwo1pjt.us-east-1.rds.amazonaws.com -u dwrecallwriter -pTsLdpZumzovrAvttIqnePCJhIVxZZ7bd <<EOF
+RENAME table rec_recall.ads_vova_kg_d to rec_recall.ads_vova_kg_d_pre,rec_recall.ads_vova_kg_d_new to rec_recall.ads_vova_kg_d;
+EOF
+echo "-------rename结束--------"
+if [ $? -ne 0 ];then
+  exit 1
+fi
+
+
+sh /mnt/vova-bigdata-scripts/common/table_check.sh  ads.ads_vova_knowledge_graph_bod_goods_rank_data 800 "pt='${pre_date}'"
+
+if [ $? -ne 0 ];then
+  exit 1
+fi
 
 sql="
 drop table if exists rec_recall.ads_vova_knowledge_graph_bod_goods_rank_data_pre;
@@ -41,6 +105,8 @@ fi
 sqoop export \
 -Dorg.apache.sqoop.export.text.dump_data_on_error=true \
 -Dsqoop.export.records.per.statement=500 \
+-Dmapreduce.map.memory.mb=8192 \
+-Dmapreduce.reduce.memory.mb=8192 \
 --connect jdbc:mysql://rec-recall.cluster-cznqgcwo1pjt.us-east-1.rds.amazonaws.com:3306/rec_recall \
 --username dwrecallwriter --password TsLdpZumzovrAvttIqnePCJhIVxZZ7bd \
 --m 1 \
